chore: move discord_utils
