+++ src/ibbtc_fee_collector.py
@@ -15,9 +15,9 @@ from src.tx_utils import get_effective_gas_price
 from src.tx_utils import get_gas_price_of_tx
 from src.tx_utils import get_priority_fee
 from src.web3_utils import confirm_transaction
-from src.utils import get_hash_from_failed_tx_error
-from src.utils import send_oracle_error_to_discord
-from src.utils import send_success_to_discord
+from src.discord_utils import get_hash_from_failed_tx_error
+from src.discord_utils import send_oracle_error_to_discord
+from src.discord_utils import send_success_to_discord
 
 FEE_THRESHOLD = 0.01  # ratio of gas cost to harvest amount we're ok with
