+++ src/oracle.py
@@ -23,9 +23,9 @@ from src.tx_utils import get_gas_price_of_tx
 from src.tx_utils import get_priority_fee
 from src.web3_utils import confirm_transaction
 from src.utils import get_abi
-from src.utils import get_hash_from_failed_tx_error
-from src.utils import send_oracle_error_to_discord
-from src.utils import send_success_to_discord
+from src.discord_utils import get_hash_from_failed_tx_error
+from src.discord_utils import send_oracle_error_to_discord
+from src.discord_utils import send_success_to_discord
 
 # push report to centralizedOracle
 REPORT_TIME_UTC = {"hour": 18, "minute": 30, "second": 0, "microsecond": 0}
