+++ src/rebaser.py
@@ -17,10 +17,10 @@ from src.tx_utils import get_effective_gas_price
 from src.tx_utils import get_gas_price_of_tx
 from src.tx_utils import get_priority_fee
 from src.web3_utils import confirm_transaction
-from src.utils import get_hash_from_failed_tx_error
+from src.discord_utils import get_hash_from_failed_tx_error
 from src.misc_utils import hours
-from src.utils import send_rebase_error_to_discord
-from src.utils import send_rebase_to_discord
+from src.discord_utils import send_rebase_error_to_discord
+from src.discord_utils import send_rebase_to_discord
 
 MAX_GAS_PRICE = int(1000e9)  # 1000 gwei
