+++ tests/test_tx_utils.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 from hexbytes import HexBytes
 
 from src.tx_utils import get_latest_base_fee
-from src.utils import get_hash_from_failed_tx_error
+from src.discord_utils import get_hash_from_failed_tx_error
 
 
 def test_get_latest_base_fee():
@@ -29,7 +29,7 @@ def test_get_latest_base_fee_hex_fee():
 
 
 def test_get_hash_from_failed_tx_error(mocker):
-    discord = mocker.patch("src.utils.send_error_to_discord")
+    discord = mocker.patch("src.discord_utils.send_error_to_discord")
     tx_hash = "0x123123"
     assert get_hash_from_failed_tx_error(
         ValueError({'data': {tx_hash: 123}}),
@@ -43,7 +43,7 @@ def test_get_hash_from_failed_tx_error_raises(mocker):
     In case something goes wrong, func just returns HexBytes(0)
     """
     mocker.patch(
-        "src.utils.send_error_to_discord",
+        "src.discord_utils.send_error_to_discord",
         side_effect=Exception,
     )
     tx_hash = "0x123123"
